+++ tests/test_visitors/test_ast/test_conditions/test_simplified_match_conditions.py
@@ -24,11 +24,28 @@ match subject:
         pass
 """
 
-simplifiable_guard_match = """
+
+single_case_sequence = """
+match subject:
+    case [1, 2, 3]:
+        pass
+"""
+
+single_case_wildcard = """
+match subject:
+    case _:
+        pass
+"""
+
+single_case_literal = """
+match subject:
+    case '1':
+        pass
+"""
+
+single_case_guard = """
 match subject:
     case x if x > 0:
-        pass
-    case _:
         pass
 """
 
@@ -91,6 +108,89 @@ match subject:
         pass
 """
 
+sequence_mapping_template = """
+match subject:
+    case {pattern}:
+        pass
+    case _:
+        pass
+"""
+
+# Wrong sequence:
+_SIMPLIFIABLE_PATTERNS = (
+    "['a', False, ['b'], {'c': 2, 'x': None}]",
+    "{'a': '1', 'b': {'2': [1]}}",
+    '[1, 2, 3]',
+    "['a', 'b', 'c']",
+    '[True, False]',
+    "[1, 'a', 2]",
+    '[True, 1, False]',
+    "[True, 'a', False]",
+    "[True, 1, 'a', None]",
+)
+
+# Correct sequence:
+_NON_SIMPLIFIABLE_PATTERNS = (
+    '[x, y, z]',
+    '{"key": x}',
+    '{"key": [x]}',
+    '{"key": 1, **kwargs}',
+    "[{'a': x}]",
+    "{'a': '1', 'b': {'2': x}}",
+    '[1, [x]]',
+    '[1, *rest]',
+)
+
+
+@pytest.mark.parametrize('pattern', _SIMPLIFIABLE_PATTERNS)
+def test_simplifiable_sequence(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    pattern,
+):
+    """Test that patterns with sequence/mapping require simplification."""
+    tree = parse_ast_tree(sequence_mapping_template.format(pattern=pattern))
+    visitor = SimplifiableMatchVisitor(default_options, tree=tree)
+    visitor.run()
+    assert_errors(visitor, [SimplifiableMatchViolation])
+
+
+@pytest.mark.parametrize('pattern', _NON_SIMPLIFIABLE_PATTERNS)
+def test_non_simplifiable_patterns(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    pattern,
+):
+    """Test that patterns do not require simplification."""
+    tree = parse_ast_tree(sequence_mapping_template.format(pattern=pattern))
+    visitor = SimplifiableMatchVisitor(default_options, tree=tree)
+    visitor.run()
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize(
+    'code',
+    [
+        single_case_sequence,
+        single_case_literal,
+        single_case_guard,
+        single_case_wildcard,
+    ],
+)
+def test_simplifiable_single_case_match(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    code,
+):
+    """Test that match with only one case."""
+    tree = parse_ast_tree(code)
+    visitor = SimplifiableMatchVisitor(default_options, tree=tree)
+    visitor.run()
+    assert_errors(visitor, [SimplifiableMatchViolation])
+
 
 @pytest.mark.parametrize(
     'code',
@@ -121,18 +221,6 @@ def test_simplifiable_single_match(
     assert_errors(visitor, [SimplifiableMatchViolation])
 
 
-def test_simplifiable_guarded_match(
-    assert_errors,
-    parse_ast_tree,
-    default_options,
-):
-    """Test that guarded irrefutable match is simplified."""
-    tree = parse_ast_tree(simplifiable_guard_match)
-    visitor = SimplifiableMatchVisitor(default_options, tree=tree)
-    visitor.run()
-    assert_errors(visitor, [SimplifiableMatchViolation])
-
-
 @pytest.mark.parametrize(
     ('left', 'right'),
     [
@@ -157,7 +245,7 @@ def test_simplifiable_union_match(
 ):
     """Test that union pattern raises violation."""
     tree = parse_ast_tree(
-        simplifiable_union_match_match.format(left, right, as_binding)
+        simplifiable_union_match_match.format(left, right, as_binding),
     )
     visitor = SimplifiableMatchVisitor(default_options, tree=tree)
     visitor.run()