+++ wemake_python_styleguide/visitors/ast/operators.py
@@ -140,7 +140,7 @@ class UselessOperatorsVisitor(base.BaseNodeVisitor):  # noqa: WPS214
             )
             if has_useless_constant:
                 self.add_violation(
-                    consistency.MeaninglessBooleanOperationViolation(node)
+                    consistency.MeaninglessBooleanOperationViolation(node),
                 )
                 return
 
@@ -163,7 +163,7 @@ class UselessOperatorsVisitor(base.BaseNodeVisitor):  # noqa: WPS214
 
             if has_useless_name:
                 self.add_violation(
-                    consistency.MeaninglessBooleanOperationViolation(node)
+                    consistency.MeaninglessBooleanOperationViolation(node),
                 )
                 return
 
@@ -213,7 +213,7 @@ class UselessOperatorsVisitor(base.BaseNodeVisitor):  # noqa: WPS214
             )
             if is_identical_constants:
                 self.add_violation(
-                    consistency.MeaninglessNumberOperationViolation(right)
+                    consistency.MeaninglessNumberOperationViolation(right),
                 )
 
     def _get_non_negative_nodes(